+++ controllers/statistics_controller.go
@@ -102,13 +102,26 @@ func GetShanghaiStatistics(w http.ResponseWriter, r *http.Request) {
 	json.NewEncoder(w).Encode(stats)
 }
 
-// GetGlobalStatistics will return some global statistics for all legs played
+// GetGlobalStatistics will return some global statistics for all matches
 func GetGlobalStatistics(w http.ResponseWriter, r *http.Request) {
 	SetHeaders(w)
 
 	global, err := data.GetGlobalStatistics()
 	if err != nil {
 		log.Println("Unable to get global statistics", err)
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	json.NewEncoder(w).Encode(global)
+}
+
+// GetGlobalStatisticsFnc will return global fish and chips counter
+func GetGlobalStatisticsFnc(w http.ResponseWriter, r *http.Request) {
+	SetHeaders(w)
+
+	global, err := data.GetGlobalStatisticsFnc()
+	if err != nil {
+		log.Println("Unable to get global fish and chips statistics", err)
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}