+++ kcapp-api.go
@@ -86,6 +86,7 @@ func main() {
 	router.HandleFunc("/statistics/shanghai/{from}/{to}", controllers.GetShanghaiStatistics).Methods("GET")
 	router.HandleFunc("/statistics/7/{from}/{to}", controllers.GetShanghaiStatistics).Methods("GET")
 	router.HandleFunc("/statistics/global", controllers.GetGlobalStatistics).Methods("GET")
+	router.HandleFunc("/statistics/global/fnc", controllers.GetGlobalStatisticsFnc).Methods("GET")
 	router.HandleFunc("/statistics/office/{from}/{to}", controllers.GetOfficeStatistics).Methods("GET")
 	router.HandleFunc("/statistics/office/{office_id}/{from}/{to}", controllers.GetOfficeStatistics).Methods("GET")
 	router.HandleFunc("/statistics/{dart}/hits", controllers.GetDartStatistics).Methods("GET")